Remove people from teams
Companies come and go, and so do their employees. Some companies in the
community have had quite the turmoil in terms of their staffing, and
most the list of employees are out of date.

To mitigate that fact, I am removing the list of employees of each
company. If you want to get in touch with someone, do so over our Chat
or go to a meetup.