Update gems
The config change is "if you run this locally, and the Jekyll template
files are being read, and interpreted as broken dates in the YAML
frontmatter, it is due to the fact that the vendor/bundle directory, or
the tmp/ directory is being scanned for Jekyll content".

- pathutil does not run right without the branch (not yet merged) which
fixes a keyword arguments issue.
- webrick is no longer included in Ruby 3, so we need it in the Gemfile.